Fix low contrast in night-mode code-blocks (#39967)
First noticed on "Divide the Grid Into an Area Template", text is hard to read due to low contrast in the css code block. Problem was discussed in issue #39967, and this change was proposed to boost contrast above the minimum acceptable ratio of 4.5:1.

Checklist:

[ x ] I have read freeCodeCamp's contribution guidelines.
[ x ] My pull request has a descriptive title (not a vague title like Update index.md)
[ x ] My pull request targets the master branch of freeCodeCamp.
[ x ] All the files I changed are in the same world language, for example: only English changes, or only Chinese changes, etc.

Closes #39967

Credit to @ojeytonwilliams for suggesting the change!